style: add skeleton to wallet
