+++ src/lib/utils.ts
@@ -24,7 +24,7 @@ export function Tovalue(v: string): string {
 	return value
 }
 
-export function CalculateDate(date: string): string{
+export function CalculateDate(date: string): string {
 	const now = new Date()
 	const diff = Number(now) - Number(new Date(date))
 
@@ -51,4 +51,18 @@ export function CalculateDate(date: string): string{
 	} else {
 		return `${years} years ago`
 	}
+}
+
+export function FormatTimeStamp(timeStamp: number):string {
+	const date = new Date(timeStamp * 1000)
+
+	const year = date.getFullYear()
+	const month = ("0" + (date.getMonth() + 1)).slice(-2)
+	const day = ("0" + date.getDate()).slice(-2)
+	const hours = ("0" + date.getHours()).slice(-2)
+	const minutes = ("0" + date.getMinutes()).slice(-2)
+	const seconds = ("0" + date.getSeconds()).slice(-2)
+
+	const formattedTime = `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`
+	return formattedTime
 }
